BUILD: Distro net-snmp-perl installs too many dependencies
On el9 and el10, net-snmp-perl depends on net-snmp-devel causing
tons of devel dependencies to be installed. We do not want devel
packages on servers so we have added a package called
net-snmp-perl-minimal in our repositories which does only
provides the perl SNMP modules without all the crazy deps that
distro package net-snmp-perl requires.